Task1 - Cloud Architecture Components Diagram
Scenario: A mid-sized e-commerce company is planning to migrate its on-premise infrastructure to leverage cloud technologies for better scalability, reliability, and service
availability capabilities. The company has a set of 3 microservices that are based on docker images deployed in one on-premise server, a relational database and a non-relational database which stores configurations as JSON records. The company expects to migrate the on-premise services into the cloud using AWS Services. The company is also keen on ensuring high service availability and horizontal scaling for the docker based service leveraging security best practices.