Task 2.1 - Build the get_data.py script
a) Get familiar with NeoWs API to retrieve data about NEOs. Generate your free API
Key and take into consideration the Daily API Rate Limits for this test.
b) Use the API Endpoint “Neo - Browse” and use the page 1755 as filter. Extract the
results.
c) Convert the API response as a JSON object and store it as a JSON file in the
output/response.json filename path.